--- src/utils/validation.ts
@@ -1,4 +1,0 @@
-export const isNotNumber = (value: string) =>
-  Number.isNaN(Number(value)) || value.includes(' ') || value.includes('.');
-
-export const isNotAlphabet = (value: string) => !/^[a-zA-Z\s]*$/.test(value);
